requirements.txt
python-telegram-bot[rate-limiter]==20.1
PyYAML
pymongo==4.3.3
python-dotenv==0.21.0
pika
numpy
minio